+++ karma.conf.js
@@ -26,7 +26,7 @@ module.exports = function(config) {
         // list of files / patterns to load in the browser
         files: [
             './dist/sdk.js',
-            './test/*.js'
+            './test/**/*.js'
         ],
 
 
@@ -39,7 +39,7 @@ module.exports = function(config) {
         // preprocess matching files before serving them to the browser
         // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
         preprocessors: {
-            './test/*.js': ['browserify'],
+            './test/**/*.js': ['browserify'],
         },
 
 
@@ -106,62 +106,62 @@ module.exports = function(config) {
 };
 
 var browsers = {
-    sl_chrome43_win7: {
-      base: 'SauceLabs',
-      browserName: 'chrome',
-      platform: 'Windows 7',
-      version: '43'
-    },
-    sl_chrome26_win7: {
-      base: 'SauceLabs',
-      browserName: 'chrome',
-      platform: 'Windows 7',
-      version: '26'
-    },
-    sl_firefox37: {
-        base: 'SauceLabs',
-        browserName: 'firefox',
-        platform: 'Windows 7',
-        version: '37.0'
-    },
-    sl_firefox4: {
-        base: 'SauceLabs',
-        browserName: 'firefox',
-        platform: 'Windows 7',
-        version: '4.0'
-    },
-    sl_ie11_win81: {
-        base: 'SauceLabs',
-        browserName: 'internet explorer',
-        platform: 'Windows 8.1',
-        version: '11.0'
-    },
-    sl_ie10_win7: {
-        base: 'SauceLabs',
-        browserName: 'internet explorer',
-        platform: 'Windows 7',
-        version: '10.0'
-    },
+    // sl_chrome43_win7: {
+    //   base: 'SauceLabs',
+    //   browserName: 'chrome',
+    //   platform: 'Windows 7',
+    //   version: '43'
+    // },
+    // sl_chrome26_win7: {
+    //   base: 'SauceLabs',
+    //   browserName: 'chrome',
+    //   platform: 'Windows 7',
+    //   version: '26'
+    // },
+    // sl_firefox37: {
+    //     base: 'SauceLabs',
+    //     browserName: 'firefox',
+    //     platform: 'Windows 7',
+    //     version: '37.0'
+    // },
+    // sl_firefox4: {
+    //     base: 'SauceLabs',
+    //     browserName: 'firefox',
+    //     platform: 'Windows 7',
+    //     version: '4.0'
+    // },
+    // sl_ie11_win81: {
+    //     base: 'SauceLabs',
+    //     browserName: 'internet explorer',
+    //     platform: 'Windows 8.1',
+    //     version: '11.0'
+    // },
+    // sl_ie10_win7: {
+    //     base: 'SauceLabs',
+    //     browserName: 'internet explorer',
+    //     platform: 'Windows 7',
+    //     version: '10.0'
+    // },
     sl_ie7_winxp: {
         base: 'SauceLabs',
         browserName: 'internet explorer',
         platform: 'windows xp',
         version: '7.0'
-    },
-    sl_iphone_ios82: {
-        base: 'SauceLabs',
-        deviceName: 'iPhone Simulator',
-        "device-orientation": 'portrait',
-        browserName: 'safari',
-        platform: 'OS X 10.10',
-        version: '8.2'
-    },
-    sl_android_44: {
-        base: 'SauceLabs',
-        deviceName: 'Android Emulator',
-        browserName: 'android',
-        "device-orientation": 'portrait',
-        platform: 'Linux',
-        version: '4.4'
-    }
+    }// ,
+    // sl_iphone_ios82: {
+    //     base: 'SauceLabs',
+    //     deviceName: 'iPhone Simulator',
+    //     "device-orientation": 'portrait',
+    //     browserName: 'safari',
+    //     platform: 'OS X 10.10',
+    //     version: '8.2'
+    // },
+    // sl_android_44: {
+    //     base: 'SauceLabs',
+    //     deviceName: 'Android Emulator',
+    //     browserName: 'android',
+    //     "device-orientation": 'portrait',
+    //     platform: 'Linux',
+    //     version: '4.4'
+    // }
 };
